chore: add stop wraapper function for HttpServer.hpp and HttpsServer.hpp
